% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Homocedasticityfun.R
\name{Homocedasticity_fun}
\alias{Homocedasticity_fun}
\title{Homocedasticity tests}
\usage{
Homocedasticity_fun(x, y)
}
\arguments{
\item{x}{A vector of numerical values}

\item{y}{A vector of groups}
}
\description{
This function calculates the p values of the following homocedasticity tests:
Bartlett
Levene
}
\examples{
x = rnorm(1000)
group = rep(c("a", "b", "c", "d", "e"), 200)
Homocedasticity_fun(x, group)
}
\keyword{Homocedasticity}

