% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ImportantMeasures.R
\name{Important_measures}
\alias{Important_measures}
\title{Other rules measures}
\usage{
Important_measures(X, X.Transactions)
}
\arguments{
\item{X}{A set of rules}

\item{X.Transactions}{The set of transactions where the rules came from}
}
\description{
This function calculates Kulczinsky and Imbalance ratio for rules generated with the apriori algorithm of the arules package
}
\examples{
Important_measures()
}
\author{
JJ
}
\keyword{Imbalance}
\keyword{Kulczynsky,}
\keyword{apriori}
\keyword{ratio,}
\keyword{rules,}

