% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Normalityfun.R
\name{Normality_fun}
\alias{Normality_fun}
\title{Normality tests}
\usage{
Normality_fun(x)
}
\arguments{
\item{x}{A vector of numerical values}
}
\description{
This function calculates the p values of the following normality tests:
Anderson-Darling
Shapiro-Wilks
Lilliefors
Cramer Von Misses
}
\examples{
x = rnorm(1000)
Normality_fun()
}
\keyword{Normality}

