% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/PCbiplot2.R
\name{PCbiplot2}
\alias{PCbiplot2}
\title{Biplot FactoMineR PCA object}
\usage{
PCbiplot2(X, x = "Dim.1", y = "Dim.2", size.obs.text = 3)
}
\arguments{
\item{X}{A FactoMineR PCA object}

\item{x}{Dimension of the PCA object for the x axis}

\item{y}{Dimension of the PCA object for the y axis}

\item{size.obs.text}{Text size for the labels}
}
\description{
Plots biplot of a FactoMineR PCA object
}
\examples{
http://stackoverflow.com/questions/22381560/how-to-create-a-biplot-with-factominer
}
\author{
http://stackoverflow.com/questions/22381560/how-to-create-a-biplot-with-factominer
PCbiplot2()
}
\keyword{FactoMineR}
\keyword{PCA,}
\keyword{ggplot,}

