% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Rules.Prunning.fun.R
\name{Rules_prunning}
\alias{Rules_prunning}
\title{Prunnig redundant rules based on lift}
\usage{
Rules_prunning(X)
}
\arguments{
\item{X}{A set of rules}
}
\description{
This function prunes the rules obtained from the apriori algorithm on the arules package based on lift
}
\examples{
http://www.rdatamining.com/examples/association-rules
}
\author{
http://www.rdatamining.com/examples/association-rules
}
\keyword{Apriori}
\keyword{Prunning}
\keyword{Rules}

